requirements.txt
gym==0.26.0
h5py==3.9.0
ipykernel==6.25.2
ipython==8.16.1
Keras==2.1.3
Markdown==3.4.4
matplotlib==3.8.0
matplotlib-inline==0.1.6
more-itertools==8.10.0
networkx==3.10
notebook==7.0.4
numba==0.59.0
numpy==1.23.2
onnx==1.12.0
Pillow==9.0.1
psutil==5.9.5
pynndescent==0.5.11
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
scikit-learn==1.3.2
scipy==1.9.0
tensorboard==2.14.1
tensorflow==2.14.0
termcolor==2.3.0
torch==2.1.0
torch-tb-profiler==0.4.3
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.1
wandb==0.15.12
zipp==1.0.0
